Reorganize view files and update routes
