+++ src/components/DegreelistDisplay.vue
@@ -1,17 +1,17 @@
 <template>
 <tr class="list-item">
-    <td>{{ degree.number }}</td>
-    <td>{{ degree.name }}</td>
+    <td>{{ degree.plan_id }}</td>
+    <td>{{ degree.plan_name }}</td>
     <td>
-        <router-link :to="{ name: 'view', params: { id: degree.id } }">
+        <router-link :to="{ name: 'view-degree', params: { id: degree.plan_id } }">
             <span>View</span>
         </router-link>
         <span> | </span>
-        <router-link :to="{ name: 'edit', params: { id: degree.id } }">
+        <router-link :to="{ name: 'edit-degree', params: { id: degree.plan_id } }">
             <span>Edit</span>
         </router-link>
         <span> | </span>
-        <router-link :to="{ name: 'delete', params: { id: degree.id } }">
+        <router-link :to="{ name: 'delete-degree', params: { id: degree.plan_id } }">
             <span>Delete</span>
         </router-link>
     </td>
