+++ src/views/course/CourseDelete.vue
