+++ src/views/course/CourseView.vue
